requirements.txt
selenium==4.12.0
beautifulsoup4==4.12.2
pandas==2.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
lxml==4.9.3
tqdm==4.66.1
